==> 01_ANALISI_ESPLORATIVA.ipynb <==
# # Test di creazione di un file.
# # Per esportarlo cliccare con il destro sul file generato e selezionare "Copy downloaded link"
# import pandas as pd
# df = pd.DataFrame([{'nome': 'fra', 'eta': 40}, {'nome': 'gio', 'eta': 50}])
# df.to_excel("./data/test.xlsx", index=False)

# CONNESSIONE A WIH

import pandas as pd
import numpy as np
from tqdm import tqdm
from pyathena import connect
from pyathena.pandas.util import as_pandas
import urllib.request as request

# install requirements
# Setup
cursor = connect(
 region_name="eu-west-1",
 s3_staging_dir="s3://wih.aws-athena-query-results.eu-west-1/",
 work_group="primary"
 ).cursor()

pd.set_option("display.max_columns", None)

# catalogo dati versionati 
cursor.execute("""
    SHOW TABLES IN AwsDataCatalog.wih_oja_versioned;
    """, {'param': 'a string'})

tables_versioned=as_pandas(cursor)
tables_versioned

cursor.execute("""
    SHOW TABLES IN AwsDataCatalog.wih_oja_latest;
    """, {'param': 'a string'})

tables_latest=as_pandas(cursor)
tables_latest

# Analisi tabella "wih_oja_blended"

## Totale righe e annunci distinti

cursor.execute("SELECT count(*), count(distinct oja_id) FROM wih_oja_latest.wih_oja_blended")
as_pandas(cursor)

## Annunci per anno

cursor.execute("""SELECT first_active_year, count(*) as n_rows, count(distinct oja_id) as n_oja
                    FROM wih_oja_latest.wih_oja_blended
                    GROUP BY first_active_year
                    ORDER BY first_active_year""")
df = as_pandas(cursor)
df.to_excel("./data/stat/oja_per_anno.xlsx", index=False)
df

## Annunci per anno-mese

- **N.B.** Al momento, sono disponibili i dati fino al **3° trimestre del 2024**

cursor.execute("""SELECT first_active_year, first_active_month, count(*) as n_rows, count(distinct oja_id) as n_oja
                    FROM wih_oja_latest.wih_oja_blended
                    GROUP BY first_active_year, first_active_month
                    ORDER BY first_active_year, first_active_month""")
df = as_pandas(cursor)
# df.to_excel("./data/stat/oja_per_anno_mese.xlsx", index=False)
df

## Annunci 2024 per country

cursor.execute("""
                    SELECT country_id, country, count(*) as n_rows, count(distinct oja_id) as n_oja
                    FROM wih_oja_latest.wih_oja_blended
                    WHERE first_active_year = 2024
                    GROUP BY country_id, country
                    ORDER BY country_id
                """)
df = as_pandas(cursor)
df.to_excel("./data/stat/oja_per_country.xlsx", index=False)
df

## Annunci 2024 per language

cursor.execute("""
                    SELECT upper(language) as language, count(*) as n_rows, count(distinct oja_id) as n_oja
                    FROM wih_oja_latest.wih_oja_blended
                    WHERE first_active_year = 2024
                    GROUP BY upper(language)
                    ORDER BY upper(language)
                """)
df = as_pandas(cursor)
df.to_excel("./data/stat/oja_per_language.xlsx", index=False)
df

## Annunci 2024 per source_category

cursor.execute("""
                    SELECT source_category, count(*) as n_rows, count(distinct oja_id) as n_oja
                    FROM wih_oja_latest.wih_oja_blended
                    WHERE first_active_year = 2024
                    GROUP BY source_category
                    ORDER BY source_category
                """)
df = as_pandas(cursor)
df.to_excel("./data/stat/oja_per_source_category.xlsx", index=False)
df

## Annunci 2024 per education

cursor.execute("""
                    SELECT education_id, education, count(*) as n_rows, count(distinct oja_id) as n_oja
                    FROM wih_oja_latest.wih_oja_blended
                    WHERE first_active_year = 2024
                    GROUP BY education_id, education
                    ORDER BY education_id
                """)
df = as_pandas(cursor)
df.to_excel("./data/stat/oja_per_education.xlsx", index=False)
df

## Annunci 2024 per source_stability

- dalla pagina WIKI: "*We recommend using all sources in the dataset for descriptive, cross-sectional analysis; and to use only stable sources when dealing with time series comparisons*"

cursor.execute("""
                    SELECT source_stability, count(*) as n_rows, count(distinct oja_id) as n_oja
                    FROM wih_oja_latest.wih_oja_blended
                    WHERE first_active_year = 2024
                    GROUP BY source_stability
                    ORDER BY source_stability
                """)
df = as_pandas(cursor)
df.to_excel("./data/stat/oja_per_source_stability.xlsx", index=False)
df

## Annunci 2024, da fonti stabili, in Italia, per laureati (ED6 e ED7)
- 1.3 milioni di record

%%time
cursor.execute("""
                    SELECT *
                    FROM wih_oja_latest.wih_oja_blended
                    WHERE first_active_year = 2024 
                        and country_id='IT' 
                        and source_stability='stable' 
                        and education_id in ('ED6', 'ED7')
                """)
df = as_pandas(cursor)
#df.to_parquet("./data/wih_oja_blended_2024_stable_it_ed67.parquet")
#df

# EXPORT TABELLE

# Lista di tabelle con un numero gestibile di record
tab_list = ['codelist_escoskill_v111'
            , 'codelist_escoskill_v111_v1'
            , 'codelist_wih_oja'
            , 'codelist_wih_oja_city'
            , 'codelist_wih_oja_skill'
            , 'codelist_wih_oja_skill_hier'
            , 'wih_oja_date'
           ]

for tab in tqdm(tab_list):
    cursor.execute(f""" SELECT * FROM wih_oja_latest.{tab} """)
    df = as_pandas(cursor)
    df.to_excel(f"./data/{tab}.xlsx", index=False)

# ESEMPI TABELLE MASTER

cursor.execute("""
                    SELECT *
                    FROM wih_oja_latest.wih_oja_blended
                    ORDER BY RAND()
                    LIMIT 10000;
                """)
df = as_pandas(cursor)
df.to_excel("./data/wih_oja_blended_SAMPLE_10k.xlsx", index=False)
df.head(3)

cursor.execute("""
                    SELECT *
                    FROM wih_oja_latest.wih_oja_skill
                    LIMIT 1000
                """)
df = as_pandas(cursor)
df.head(3)

cursor.execute("""
                    SELECT oja_id
                        , first_active_date
                        , first_active_month
                        , first_active_year
                        , last_active_date
                        , language
                        , occupation4d_id
                        , occupation3d_id
                        , skill_id
                        , skill_hier3_id
                        , skill_hier2_id
                        , skill_hier1_id
                        , skill_hier0_id
                        , city_id
                        , nuts3_id
                        , country_id
                        , contract_id
                        , education_id
                        , economic_activity2d_id
                        , economic_activity1d_id
                        , activity_sector_id
                        , salary_id
                        , working_time_id
                        , experience_id
                        , source_id
                        , source_category_id
                    FROM wih_oja_latest.wih_oja_main
                    LIMIT 1000
                """)
df = as_pandas(cursor)
df.head(3)
