==> 02.2_SQL_OFFERTA_LUISS.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
from pyathena import connect
from pyathena.pandas.util import as_pandas
import urllib.request as request

cursor = connect(
 region_name="eu-west-1",
 s3_staging_dir="s3://wih.aws-athena-query-results.eu-west-1/",
 work_group="primary"
 ).cursor()

pd.set_option("display.max_columns", None)

# Selezione OJA in target
Ciclo per tutti i codici ISCO a 3 digit

Perimetro considerato:
- ITALIA
- dal 3° trimestre 2022 in poi
- Livello di istruzione almeno laurea
- sorgenti dati stabili

# Offerta formativa Luiss
offerta_liv4 = pd.read_parquet("./data/offerta_luiss/offerta_formativa_isco4.parquet")
occupations4d = offerta_liv4[['occupation4d_id']].drop_duplicates()
occupations4d.shape[0]

# Calcolo Skill per occupazione

%%time

def skill_per_professione_oc_quarter(anno, country_it, out_parquet):

    if country_it==True:
        cnty = "AND country='ITALIA'"
        territory = 'IT'
    else:
        cnty = ""
        territory = 'EU'

    for index, row in tqdm(occupations4d[['occupation4d_id']].drop_duplicates().iterrows()):
        ocse4 = row['occupation4d_id']
    
        cursor.execute(f"""
                SELECT '{territory}' as territory
                    , first_active_year
                    , CASE 
                        WHEN first_active_month >= 1 AND first_active_month <= 3 THEN 1
                        WHEN first_active_month >= 4 AND first_active_month <= 6 THEN 2
                        WHEN first_active_month >= 7 AND first_active_month <= 9 THEN 3
                        WHEN first_active_month >= 10 AND first_active_month <= 12 THEN 4
                      END as quarter
                    , occupation4d_id
                    , skill_id
                    , COUNT(DISTINCT oja_id) as n_oja
                FROM wih_oja_latest.wih_oja_blended
                WHERE source_stability='stable'
                    AND education_id IN ('ED6', 'ED7', 'ED8') -- I missing sono 65 su 170k, quindi ok!
                    {cnty}
                    AND first_active_year={anno}
                    AND occupation4d_id='{ocse4}'
                GROUP BY '{territory}'
                    , first_active_year
                    , CASE 
                        WHEN first_active_month >= 1 AND first_active_month <= 3 THEN 1
                        WHEN first_active_month >= 4 AND first_active_month <= 6 THEN 2
                        WHEN first_active_month >= 7 AND first_active_month <= 9 THEN 3
                        WHEN first_active_month >= 10 AND first_active_month <= 12 THEN 4
                      END as quarter
                    , occupation4d_id
                    , skill_id
                ;
        """)
    
        if index==0:
            skill_q = as_pandas(cursor)
        else:
            skill_q = pd.concat([skill_q, as_pandas(cursor)])
    
    
    skill_q.to_parquet(out_parquet)

# IT: 2024 - T4
skill_per_professione_oc_quarter(anno=2024, country_it=True, out_parquet=f"./data/report/IT_skill_per_occupation_2024_T.parquet")
# EU: 2024 - T4
skill_per_professione_oc_quarter(anno=2024, country_it=False, out_parquet=f"./data/report/EU_skill_per_occupation_2024_T.parquet")

# IT: 2023 - T4
skill_per_professione_oc_quarter(anno=2023, country_it=True, out_parquet=f"./data/report/IT_skill_per_occupation_2023_T.parquet")
# EU: 2023 - T4
skill_per_professione_oc_quarter(anno=2023, country_it=False, out_parquet=f"./data/report/EU_skill_per_occupation_2023_T.parquet")

# Calcolo oja per occupazione

def oja_per_professione_oc_quarter(anno, country_it, out_parquet):

    if country_it==True:
        cnty = "AND country='ITALIA'"
        territory = 'IT'
    else:
        cnty = ""
        territory = 'EU'

    for index, row in tqdm(occupations4d[['occupation4d_id']].drop_duplicates().iterrows()):
        ocse4 = row['occupation4d_id']
    
        cursor.execute(f"""
                        SELECT '{territory}' as territory
                            , first_active_year
                            , CASE 
                                WHEN first_active_month >= 1 AND first_active_month <= 3 THEN 1
                                WHEN first_active_month >= 4 AND first_active_month <= 6 THEN 2
                                WHEN first_active_month >= 7 AND first_active_month <= 9 THEN 3
                                WHEN first_active_month >= 10 AND first_active_month <= 12 THEN 4
                              END as quarter
                            , occupation4d_id
                            , COUNT(DISTINCT oja_id) as n_oja
                        FROM wih_oja_latest.wih_oja_blended
                        WHERE source_stability = 'stable'
                            AND education_id IN ('ED6', 'ED7', 'ED8')
                            {cnty}
                            AND first_active_year = {anno}
                            AND first_active_month >= {month_start} AND first_active_month <= {month_end}
                            AND occupation4d_id = '{ocse4}'
                        GROUP BY '{territory}'
                            , first_active_year
                            , CASE 
                                WHEN first_active_month >= 1 AND first_active_month <= 3 THEN 1
                                WHEN first_active_month >= 4 AND first_active_month <= 6 THEN 2
                                WHEN first_active_month >= 7 AND first_active_month <= 9 THEN 3
                                WHEN first_active_month >= 10 AND first_active_month <= 12 THEN 4
                              END
                            , occupation4d_id
                    """)
    
        if index==0:
            skill_q = as_pandas(cursor)
        else:
            skill_q = pd.concat([skill_q, as_pandas(cursor)])
    
    
    skill_q.to_parquet(out_parquet)


# IT: 2024
oja_per_professione_oc_quarter(anno=2024, country_it=True, out_parquet=f"./data/report/IT_oja_per_occupation_2024_T.parquet")
# EU: 2024
oja_per_professione_oc_quarter(anno=2024, country_it=False, out_parquet=f"./data/report/EU_oja_per_occupation_2024_T.parquet")

# IT: 2023
oja_per_professione_oc_quarter(anno=2023, country_it=True, out_parquet=f"./data/report/IT_oja_per_occupation_2023_T.parquet")
# EU: 2023
oja_per_professione_oc_quarter(anno=2023, country_it=False, out_parquet=f"./data/report/EU_oja_per_occupation_2023_T.parquet")
